backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
motor==3.7.1
email-validator==2.2.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
slowapi==0.1.9
